Chore: Move lib/file-finder.js to lib/util/ (refs #10559)
