+++ lib/util/file-finder.js
